+++ src/renderers/metal/main.mm
@@ -131,10 +131,10 @@ int app_main(int, char**) {
             id<MTLCommandBuffer> commandBuffer = [commandQueue commandBuffer];
 
             renderPassDescriptor.colorAttachments[0].clearColor = MTLClearColorMake(
-                clear_color[0] * clear_color[3],
-                clear_color[1] * clear_color[3],
-                clear_color[2] * clear_color[3],
-                clear_color[3]
+                static_cast<double>(clear_color[0] * clear_color[3]),
+                static_cast<double>(clear_color[1] * clear_color[3]),
+                static_cast<double>(clear_color[2] * clear_color[3]),
+                static_cast<double>(clear_color[3])
             );
             renderPassDescriptor.colorAttachments[0].texture = drawable.texture;
             renderPassDescriptor.colorAttachments[0].loadAction = MTLLoadActionClear;
